Add files via upload
This is a Business Intelligence analysis of Oracle Support Team in terms of performance per member. From a decision making point of vue,  it can be interesting to have a good knowledge of your team and take the right decision. Therefore, we directed our analysis in terms of team performance.